Switched storage vertex bitangent sign to uint8_t
It doesn't really help with compressed file size and it doesn't help
at all before I implement separate streams so maybe it was a stupid idea.

+++ source/core/scene/shape/triangle/bvh/triangle_bvh_indexed_data.hpp
@@ -83,7 +83,7 @@ class Indexed_data {
     size_t num_bytes() const noexcept;
 
     struct alignas(16) Index_triangle {
-        Index_triangle(uint32_t a, uint32_t b, uint32_t c, float bitangent_sign,
+        Index_triangle(uint32_t a, uint32_t b, uint32_t c, uint8_t bitangent_sign,
                        uint32_t material_index) noexcept;
 
         uint32_t a, b, c;
@@ -144,7 +144,7 @@ class Indexed_data_interleaved {
     uint32_t num_vertices_;
 
     struct alignas(16) Index_triangle {
-        Index_triangle(uint32_t a, uint32_t b, uint32_t c, float bitangent_sign,
+        Index_triangle(uint32_t a, uint32_t b, uint32_t c, uint8_t bitangent_sign,
                        uint32_t material_index);
 
         uint32_t a, b, c;
@@ -210,7 +210,7 @@ class Hybrid_data {
     size_t num_bytes() const;
 
     struct alignas(16) Index_triangle {
-        Index_triangle(uint32_t a, uint32_t b, uint32_t c, float bitangent_sign,
+        Index_triangle(uint32_t a, uint32_t b, uint32_t c, uint8_t bitangent_sign,
                        uint32_t material_index);
 
         uint32_t a, b, c;
